% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emld-methods.R
\name{as_json}
\alias{as_json}
\title{as_json}
\usage{
as_json(x, file = NULL)
}
\arguments{
\item{x}{an emld object}

\item{file}{optional path to write out to file.
Otherwise, defaults to NULL and will return a json object.}
}
\description{
as_json
}
