% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emld-methods.R
\name{as_xml}
\alias{as_xml}
\title{as_xml}
\usage{
as_xml(x, file = NULL)
}
\arguments{
\item{x}{an emld object}

\item{file}{optional path to write out to file.
Otherwise, defaults to NULL and will return an xml_document object.}
}
\description{
as_xml
}
