% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_version.R
\name{ask_eml_version}
\alias{ask_eml_version}
\title{Ask the user to choose an EML version}
\usage{
ask_eml_version()
}
\value{
An EML version string suitable for \code{options(emld_db)}
}
\description{
Ask the user to choose an EML version
}
