% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_version.R
\name{eml_ns}
\alias{eml_ns}
\title{Get the XML namespace for a version of EML}
\usage{
eml_ns(version = eml_version())
}
\arguments{
\item{version}{EML version, currently either eml-2.2.0 (current version) or
eml-2.1.1. Defaults to current version.}
}
\value{
returns the full XML namespace URI for the specified version of the
schema
}
\description{
Utility function for use when filling in \code{xmlns}, \code{schemaLocation}, or
\code{vocab} in various representations of EML. This is a little more future-proof
than keeping a dictionary for each version since this won't break on the next
release.
}
