% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_version.R
\name{eml_schema_location}
\alias{eml_schema_location}
\title{Get an appropriate \code{schemaLocation} value for a given version of the schema}
\usage{
eml_schema_location(version = eml_version(), url = NULL)
}
\arguments{
\item{version}{Optional. Override the version of the schema.}

\item{url}{Override the URL portion of the \code{schemaLocation} value.}
}
\value{
Returns a string suitable as a value for \code{schemaLocation}.
}
\description{
This is a simple helper to make filling in the \code{schemaLocation} attribute
on documents this package creates.
}
\examples{
\dontrun{
# Get an appropriate schemaLocation value for the current version fo EML
eml_schema_location()

# Get an appropriate value for EML 2.1.1
eml_schema_location("eml-2.1.1")

# Set `schemaLocation` to a custom URL of your choosing
eml_schema_location(url = "https://example.org/eml.xsd")
}
}
