% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_version.R
\name{eml_version}
\alias{eml_version}
\title{Set or check the EML version default}
\usage{
eml_version(version = getOption("emld_db", "eml-2.1.1"))
}
\arguments{
\item{version}{EML version, currently either eml-2.2.0 (release candidate),
or eml-2.1.1 (current version, 2011 - 2019)}
}
\value{
returns the EML version string. As a side-effect, sets the
requested version as the default version by setting the \code{emld_db}
variable in \code{\link[=options]{options()}}.
}
\description{
Set or check the EML version default
}
\examples{
eml_version()

}
