% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_validate.R
\name{find_real_root_name}
\alias{find_real_root_name}
\title{Get the real \code{QName} for the root element}
\usage{
find_real_root_name(doc)
}
\arguments{
\item{doc}{An \code{xml_document}}
}
\value{
A \code{list} with elements \code{prefix} and \code{name}. \code{prefix} will be \code{NULL}
if the element has no namespace prefix but \code{name} will always be a
\code{character}.
}
\description{
If the root element uses a prefix (e.g., foo:bar), \code{xml2} doesn't support
getting that prefix after the document has been read into R though it does
retain it internally which is good.
}
