% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_validate.R
\name{guess_root_schema}
\alias{guess_root_schema}
\title{Find the root schema module and version}
\usage{
guess_root_schema(doc)
}
\arguments{
\item{doc}{An \code{xml_document}}
}
\value{
If found, a list with names 'version', 'module', and `namespace. If
not found, throws an error.
}
\description{
This is a workaround for a limitation in the \code{xml2} package where it's not
possible to get the QName of an XML document. For example, if you read in a
document that has a root of https://example.com/foo but is prefixed as
ex:foo, \code{xml2::get_name()} on the root only returns \code{foo} and not \code{ex:foo} or
\verb{https://example.com/foo} which would be super helpful here.
}
\details{
Instead, we assume we're talking about an EML document and scan the list of
defined namespaces for a pattern like {scheme}://{host}/{module}-{version}
and match when the name (local) of the root is found at the {module}
position.

This is rigid and a bit hacky but it should work.
}
