% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emljson.R
\name{parse_eml}
\alias{parse_eml}
\title{parse_eml}
\usage{
parse_eml(x, drop_comment = TRUE, add_context = FALSE)
}
\arguments{
\item{x}{path to an EML file}
}
\description{
Parse an EML file into an R list object
}
