% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{template}
\alias{template}
\title{template}
\usage{
template(object, attributes = FALSE)
}
\arguments{
\item{object}{the name of an eml object to create}

\item{attributes}{logical, default FALSE. Should template include
properties that are represented as XML attributes in EML? (e.g. system, id. etc)}
}
\value{
a list with elements named according to the properties of the object.
        This can be coerced into EML, see vignettes. NULL-valued elements (~)
        can take a data entry directly, while empty list()-valued elements ([])
        indicate properties that take other eml objects as values.
}
\description{
template
}
\examples{
template("creator")

}
