% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emljson.R
\name{xml_to_json}
\alias{xml_to_json}
\title{xml_to_json}
\usage{
xml_to_json(x, out = NULL)
}
\arguments{
\item{x}{path to an EML file}

\item{out}{output filename; optional, if ommitted will return JSON string}
}
\value{
json string or json output file
}
\description{
Convert an EML file into a json file/string
}
\examples{

ex <- system.file("extdata/example.xml", package = "emljson")
xml_to_json(ex)

}
