% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{template}
\alias{template}
\title{Create a template for an EML object}
\usage{
template(object)
}
\arguments{
\item{object}{the name of an eml object to create}
}
\value{
a list with elements named according to the properties of the object.
This can be coerced into EML, see vignettes. NULL-valued elements (~)
can take a data entry directly, while empty list()-valued elements ({})
indicate properties that take other eml objects as values.
}
\description{
Create a template for an EML object
}
\details{
Note: while this function can be called in recursions, doing so may be a bad idea.
}
\examples{
template("creator")
}
